Move Localizer to localization package
