+++ internal/renderer/renderer_j1.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 
 	"github.com/chubin/wttr.in/internal/domain"
-	"github.com/chubin/wttr.in/internal/weather"
+	"github.com/chubin/wttr.in/internal/localization"
 )
 
 type J1Renderer struct{}
 
-func (r *J1Renderer) Render(query domain.Query, localizer weather.Localizer) (domain.RenderOutput, error) {
+func (r *J1Renderer) Render(query domain.Query, localizer localization.Localizer) (domain.RenderOutput, error) {
 	if query.Weather == nil {
 		return domain.RenderOutput{}, fmt.Errorf("no weather data provided")
 	}
