+++ internal/renderer/v1/renderer.go
@@ -14,7 +14,6 @@ import (
 	"github.com/chubin/wttr.in/internal/domain"
 	"github.com/chubin/wttr.in/internal/localization"
 	"github.com/chubin/wttr.in/internal/options"
-	w "github.com/chubin/wttr.in/internal/weather"
 )
 
 // V1Renderer renders weather in the classic wttr.in v1 style.
@@ -31,7 +30,7 @@ func NewV1Renderer() *V1Renderer {
 }
 
 // Render implements the weather.Renderer interface.
-func (r *V1Renderer) Render(query domain.Query, localizer w.Localizer) (domain.RenderOutput, error) {
+func (r *V1Renderer) Render(query domain.Query, localizer localization.Localizer) (domain.RenderOutput, error) {
 	if query.Weather == nil || len(*query.Weather) == 0 {
 		return domain.RenderOutput{}, errors.New("no weather data provided")
 	}
