+++ internal/renderer/v2/renderer.go
@@ -11,7 +11,6 @@ import (
 	"github.com/chubin/wttr.in/internal/domain"
 	"github.com/chubin/wttr.in/internal/localization"
 	"github.com/chubin/wttr.in/internal/options"
-	w "github.com/chubin/wttr.in/internal/weather"
 )
 
 // V2Renderer implements the rich panel view (temperature diagram, rain sparkline,
@@ -26,7 +25,7 @@ func NewV2Renderer() *V2Renderer {
 type Localize func(text string) string
 
 // Render converts the query's weather JSON into the v2 terminal weather report.
-func (r *V2Renderer) Render(q domain.Query, localizer w.Localizer) (domain.RenderOutput, error) {
+func (r *V2Renderer) Render(q domain.Query, localizer localization.Localizer) (domain.RenderOutput, error) {
 	if q.Weather == nil || len(*q.Weather) == 0 {
 		return domain.RenderOutput{}, fmt.Errorf("no weather data available")
 	}
